+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.16.2
+
+### Fix
+
+- WorkflowRun : Compute memory usage, subtract before_memory
+
+
 ## 0.16.1
 
 ### Fix